Refactoreret færdig
