+++ src/main/java/org/example/oenskelisten/Interface/IRepository.java
@@ -7,7 +7,6 @@ import java.util.List;
 public interface IRepository<T> {
     List<T> getAll();
     T getById(int id);
-    boolean edit(T newObject);
     void add(T newObject);
     void deleteById(int id);
 }
