+++ src/main/java/org/example/oenskelisten/Repository/WishListRepository.java
@@ -1,10 +1,8 @@
 package org.example.oenskelisten.Repository;
 
 import org.example.oenskelisten.Interface.IWishListRepository;
-import org.example.oenskelisten.Model.Wish;
 import org.example.oenskelisten.Model.WishList;
 import org.example.oenskelisten.Model.WishListRowMapper;
-import org.example.oenskelisten.Model.WishRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Repository;
@@ -22,81 +20,7 @@ public class WishListRepository implements IWishListRepository {
     }
 
     @Override
-    public List<Wish> getAll() {
-        System.out.println("hello from getall");
-        //String sql ="SELECT name, description, productLink, imageLink, price FROM wishes";
-        String sql = "SELECT * FROM wishes";
-        return jdbcTemplate.query(sql, new WishRowMapper());
-        //return null;
-    }
-
-    @Override
-    public Wish getById(int id) {
-       String sql = "SELECT * FROM wishes WHERE wishID = ?";
-        try {
-            return jdbcTemplate.queryForObject(sql, new WishRowMapper(), id);
-        } catch (Exception e){
-            return null;
-        }
-    }
-
-    @Override
-    @Transactional
-    public void add(Wish wish) {
-        String sql = "INSERT INTO wishes (name, description, productlink, imagelink, price, wishlistID, reserved, reserveeID ) VALUES (?,?,?,?,?,?,?,?)";
-        jdbcTemplate.update(sql,
-                wish.getName(),
-                wish.getDescription(),
-                wish.getProductLink(),
-                wish.getImageLink(),
-                wish.getPrice(),
-                wish.getWishlistID(),
-                wish.isReserved(),
-                wish.getReserveeIDAsString());
-
-    }
-
-    @Override
-    @Transactional
-    public boolean edit(Wish wish) {
-        String sql = "UPDATE wishes SET name = ?, description = ?, productlink = ?, imagelink = ?, price = ? " +
-                "WHERE wishID = ?";
-        jdbcTemplate.update(sql, wish.getName(),
-                              wish.getDescription(),
-                              wish.getProductLink(),
-                              wish.getImageLink(),
-                              wish.getPrice(),
-                                wish.getId());
-        return true;
-    }
-
-    @Override
-    public Wish getByName(String name) {
-        String sql = "SELECT * FROM wish WHERE name = ?";
-        try {
-            return jdbcTemplate.queryForObject(sql, new WishRowMapper(), name);
-        } catch (Exception e){
-            return null;
-        }
-    }
-
-    @Override
-    public List<Wish> getWishListById(int id) {
-        String sql = "SELECT * FROM wishes WHERE wishes.wishListID = ?";
-
-        return jdbcTemplate.query(sql, new WishRowMapper(), id);
-    }
-
-    @Override
-    public void deleteById(int id) {
-        String sql ="DELETE FROM wishes WHERE wishID = ?";
-        jdbcTemplate.update(sql, id);
-
-    }
-
-    //---------------WISHLIST METHODS----------------------
-    @Override
-    public List<WishList> getAllWishLists(){
+    public List<WishList> getAll(){
         String sql = "SELECT * FROM wishlist";
         ArrayList allWishLists = new ArrayList<WishList>();
 
@@ -112,8 +36,7 @@ public class WishListRepository implements IWishListRepository {
     }
 
     @Override
-
-    public WishList getWishListModelByID(int id){
+    public WishList getById(int id){
         String sql = "SELECT \n" +
                 "    wishlist.name AS wishlist_name, \n" +
                 "    wishes.name AS wish_name, \n" +
@@ -125,5 +48,19 @@ public class WishListRepository implements IWishListRepository {
                 "WHERE wishlist.wishlistID = ?;";
         return jdbcTemplate.queryForObject(sql, new WishListRowMapper(), id);
 
+    }
+
+    @Transactional
+    @Override
+    public void add(WishList wishList) {
+        String sql = "INSERT INTO wishList(name, userID) VALUES (?,?)";
+        jdbcTemplate.update(sql, wishList.getName(), wishList.getUserID());
+    }
+
+    @Transactional
+    @Override
+    public void deleteById(int wishListID) {
+        String sql = "DELETE FROM wishList WHERE wishListID = ?";
+        jdbcTemplate.update(sql, wishListID);
     }
 }
