+++ src/main/java/org/example/oenskelisten/Service/WishListService.java
@@ -1,9 +1,9 @@
 package org.example.oenskelisten.Service;
 
 import org.example.oenskelisten.Interface.IWishListRepository;
+import org.example.oenskelisten.Interface.IWishRepository;
 import org.example.oenskelisten.Model.Wish;
 import org.example.oenskelisten.Model.WishList;
-import org.example.oenskelisten.Repository.WishListRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,49 +11,42 @@ import java.util.List;
 @Service
 public class WishListService {
     private final IWishListRepository wishListRepository;
+    private final IWishRepository wishRepository;
 
-    public WishListService(IWishListRepository repository) {
+    public WishListService(IWishListRepository repository, IWishRepository wishRepository) {
         this.wishListRepository = repository;
-    }
-
-    public List<Wish> getAllWishListItems() {
-        return wishListRepository.getAll();
-    }
-
-    public Wish getWishById(int id) {
-        return wishListRepository.getById(id);
-    }
-
-    public void addWish(Wish wish) {
-        wishListRepository.add(wish);
-    }
-
-    public void updateWish(Wish wish) {
-        wishListRepository.edit(wish);
-    }
-
-    public Wish getByName(String name) {
-        return wishListRepository.getByName(name);
-    }
-
-    public void deleteWishById(int id) {
-        wishListRepository.deleteById(id);
-    }
-
-    public List<WishList> getAllWishlists(){
-        return wishListRepository.getAllWishLists();
+        this.wishRepository = wishRepository;
     }
 
     public WishList getWishListModelByID(int id){
-        return wishListRepository.getWishListModelByID(id);
+        return wishListRepository.getById(id);
     }
 
+    // ------------------ WISH ----------------------
+    public void addWish(Wish wish) {
+        wishRepository.add(wish);
+    }
 
     public List<Wish> getWishListByID(int id) {
-        return wishListRepository.getWishListById(id);
+        return wishRepository.getWishListById(id);
     }
 
-    public void deleteWishList(int wishListID) {
-        wishListRepository.deleteById(wishListID);
+    public List<Wish> getAllWishListItems() {
+        return wishRepository.getAll();
+    }
+
+    public Wish getWishById(int id) {
+        return wishRepository.getById(id);
+    }
+
+    public void updateWish(Wish wish) {
+        wishRepository.edit(wish);
+    }
+    public void deleteWishById(int id) {
+        wishRepository.deleteById(id);
+    }
+
+    public Wish getByName(String name) {
+        return wishRepository.getByName(name);
     }
 }
